% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{BuildKmodels}
\alias{BuildKmodels}
\title{blah}
\usage{
BuildKmodels(kmdata = NULL, template = NULL, basedir = "tmp")
}
\arguments{
\item{template}{blah}

\item{basedir}{The path to the base directory of a simulation project.  The default is a subdirectory of the current directory called "tmp".  This is necessary for saving the results.}

\item{data}{blah}

\item{hmvars}{blah}
}
\value{
blah
}
\description{
blah
}
\details{
blah
}
