% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{CalcErrors}
\alias{CalcErrors}
\title{Calculate the differences between the base case and the simulation runs}
\usage{
CalcErrors(long = NULL, basecase = NULL, basedir = "tmp")
}
\arguments{
\item{long}{A long format data frame with historical data and the simulation results.}

\item{basecase}{The case against which all others will be compared.}

\item{basedir}{The path to the base directory of a simulation project.  The default is a subdirectory of the current directory called "tmp".}
}
\value{
This returns a long format data frame with production data and the calculated errors.
}
\description{
This calculates the difference between a base case and a simulation case, as well as the fractional difference.
}
\details{
First it is necessary to define the timesteps at which errors are going to be estimated.   The default dates will be those of the base case, and all other case values will need to be available at those dates (or interpolated to those dates, eventually).

The difference between a case and the base will be called ERR. The ERR divided by the base value will be called FRAC_ERR.
}
