% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gathering.R
\name{eclsum}
\alias{eclsum}
\title{Eclipse style summary file to csv}
\usage{
eclsum(sumdir = ".", outdir = NULL, projdir = NULL)
}
\arguments{
\item{outdir}{The  path to where the csv file with the summary data should be written, assumed to be the same as the indir.}

\item{projdir}{The  path to where the csv file with summary data for (possibly) multiple wells should be written, assumed to be the same as the indir.}

\item{indir}{The path to an Eclipse style simulation summary output}
}
\value{
The function writes out a csv file, and returns a dataframe with summary data.
}
\description{
This function converts Eclipse style xy type output to a csv file for later use.
}
