% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hmsetup.R
\name{expdes}
\alias{expdes}
\title{Create an experimental design using a template object.}
\usage{
expdes(obj, ...)
}
\arguments{
\item{obj}{The name of an hmvars object.}

\item{type}{The type of experimental design.  Folded Plackett-Burman is the default (and all that is supported, so far.)}

\item{basedir}{The path to the base directory of a simulation project.  The default is the current directory.  This is used to store the edited object.}
}
\value{
Returns a template object, and writes out a template object file.
}
\description{
This function reads parameters in the hmvars object as a function of variable name, and uses them in constructing an experimental design.
}
\details{
I'm thinking about it.
}
