% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_gathering.R
\name{findDecks}
\alias{findDecks}
\title{Find a set of Eclipse style input data decks}
\usage{
findDecks(basedir = NULL, casename = "^.+", ext = c("\\\\.data$",
  "\\\\.DATA$"), recursive = TRUE)
}
\arguments{
\item{basedir}{The path to the base directory of a simulation project.  The default is the current directory.}

\item{casename}{The deck basename of an Eclipse style simulation summary output.  A perl style regular expression may be used to find multiple summary files.}

\item{ext}{A list file extents used to create a search pattern.  Default is c(".data", ".DATA")}

\item{recursive}{Should we look in the directories recursively?  Default is TRUE.}
}
\value{
The function returns a list of .DATA files with the full path.
}
\description{
The function returns a list of .DATA files with the full path.
}
\details{
Default behavior to to search recursively in the basedir to find a list of ECl style input decks files.
}
