% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{ImportHist}
\alias{ImportHist}
\title{blah}
\usage{
ImportHist(ssheetfile = NULL, basedir = ".")
}
\arguments{
\item{ssheetfile}{blah}

\item{basedir}{blah}
}
\value{
blah
}
\description{
blah
}
\details{
blah
}
