% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plotting.R
\name{plotcase}
\alias{plotcase}
\title{Plot simulation results from csv files}
\usage{
plotcase(casedata)
}
\arguments{
\item{basefile}{The full path (relative or absolute) to a csv file with data to be used to compare against multiple cases, e.g the history file to be compared with multiple model runs.}

\item{compfiles}{A list of csv files with data to be compared against the base file.}

\item{template}{A file describing which data is tobe used in the comparison plots}

\item{outdir}{Where the output plots should be stored.}
}
\description{
This function accepts multiple csv files as input, and plots history match plots using ggplot
}
