% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{rss3d}
\alias{rss3d}
\title{A substitute for the DiceDesign::rss3d to fix a bug}
\usage{
rss3d(design, lower, upper, gof.test.type = "greenwood",
  gof.test.stat = NULL, transform = NULL, n.angle = 60, graphics = 1,
  trace = TRUE)
}
