% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOPM.R
\docType{package}
\name{runOPM}
\alias{runOPM}
\alias{runOPM-package}
\title{runOPM:  A package to assist in using the Open Porous Media reservoir modeling tools.}
\description{
blah blah blah
}
\author{
George Williams <gerwathome@gmail.com>
}
