% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hmsetup.R
\name{BuildDecks}
\alias{BuildDecks}
\alias{BuildDecks.hmvars}
\title{Create a set of decks using a deck template and an experimental design.}
\usage{
BuildDecks(obj, template, basedir = ".", overwrite = FALSE, cases = NULL)

\method{BuildDecks}{hmvars}(obj, template, basedir = ".", overwrite = FALSE,
  cases = NULL)
}
\arguments{
\item{obj}{The name of an hmvars object.}

\item{template}{The deck template associated with the hmvars object.}

\item{basedir}{The path to the base directory of a simulation project.  The default is the current directory.}

\item{overwrite}{TRUE means overwrite all; FALSE means don't overwrite anything; NULL (the defaults) means overwrite decks older than the hmvars object.}

\item{cases}{The default is NULL, meaning create a deck for all of the cases in the experimental design.  Alternatively, a list of line numbers for the desired cases may be submitted.}
}
\value{
Returns a list of deck files suitable for use in submitting the cases
}
\description{
This function reads an experimental design from a hmvars object and combines it with the appropriate deck template to create a series of decks implementing the experimental design.
}
\details{
I'm thinking about it.
}
