% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{BuildKModels}
\alias{BuildKModels}
\title{BuildKModels:  Build kriged proxy models}
\usage{
BuildKModels(hmvars = NULL, member_error = NULL, model_selection = NULL,
  basedir = "tmp")
}
\arguments{
\item{hmvars}{The hmvars object has the experimental design that was used to create the simulation decks.}

\item{member_error}{This contains errors calculated from the results of the simulation runs for each design point (i.e. each row) of the experimental design, and is the output of the ErrorByMember function.}

\item{model_selection}{This is output by the SelectModels function, and contains the filters to be used on the member error data to select the appropriate data for each kriged model.}

\item{basedir}{The path to the base directory of a simulation project.  The default is a subdirectory of the current directory called "tmp".  This is necessary for saving the results.}
}
\value{
Returns a list of kriged models, one model for each of the elements selected in the model selection object
}
\description{
This function creates a kriged model from the experimental design, the simulation results, and the chosen response variables.
}
\details{
blah
}
