% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{CalcErrors}
\alias{CalcErrors}
\title{Calculate the differences between the base case and the simulation runs}
\usage{
CalcErrors(long = NULL, base_case = NULL, basedir = ".")
}
\arguments{
\item{long}{A long format data frame with historical data and the simulation results.}

\item{base_case}{The case against which all others will be compared.}

\item{basedir}{The path to the base directory of a simulation project.  The default is the current directory.}
}
\value{
This returns a long format data frame with production data and the calculated errors.
}
\description{
This calculates the difference between a base case and a simulation case, as well as the fractional difference.
}
\details{
There are numerous ways to judge how well the model fits the data.  We are going to assume that a good solution has four characteristics:  first, the mean normalized absolute error is minimized; second, the mean normalized error is minimized (i.e. close to zero); third, the errors are normally distributed; and fourth, the errors have no trend with time (i.e the slope of a fit through the residuals as a function of time is near zero).  The intent of these criteria is to ensure that the quality of the fit to the historical data is comparable to the expected error in a forecast.

First it is necessary to define the timesteps at which errors are going to be estimated.   The default dates will be those of the base case, and all other case values will need to be available at those dates (or interpolated to those dates, eventually).

The difference between a case and the base will be called ERR. The ERR divided by the base value will be called FRAC_ERR.
}
