% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hmsetup.R
\name{EditVar}
\alias{EditVar}
\alias{EditVar.hmvars}
\title{Edit parameters in a template object.}
\usage{
EditVar(obj = NULL, pattern = NULL, basedir = ".",
  objname = deparse(substitute(obj)), ...)

\method{EditVar}{hmvars}(obj = NULL, pattern = NULL, basedir = ".",
  objname = deparse(substitute(obj)), ...)
}
\arguments{
\item{obj}{The name of an hmvars object.}

\item{pattern}{A perl compatible regular expression identifying the variable name for which parmeters are being edited.}

\item{basedir}{The path to the base directory of a simulation project.  The default is the current directory.  This is used to store the edited object.}

\item{objname}{This is used to save the edited object to a file. It defaults to the name of the hmvars object.}

\item{...}{A series of param=value pairs that define how the parameter values associated with a particular variable identified by "pattern" are to be set.}
}
\value{
Returns a template object, and writes out a template object file.
}
\description{
This function allows editing of parameters in the hmvars object as a function of variable name.
}
\details{
I'm thinking about it.
}
