% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{ImportHist}
\alias{ImportHist}
\title{Import production history}
\usage{
ImportHist(ssheetfile = NULL, basedir = "tmp", format = "\%d-\%b-\%Y")
}
\arguments{
\item{ssheetfile}{The full path to a .csv, .xls, or .xlsx file containing production history data.}

\item{basedir}{The path to the base directory of a simulation project.  The default is a subdirectory of the current directory called "tmp".  This is necessary for reading and writing the PROJSUm.csv file, where the history and simulation data is stored.}

\item{format}{The date format in the csv file being imported.}
}
\value{
A long format data frame with production data.
}
\description{
Import production history into a long format data frame comparable to the way simulation output is stored.
}
\details{
The spreadsheet file should include a single sheet that starts with a DATE column, followed by desired data. The DATE should be in "%d-%b-%Y" format, e.g. 01-Jan-2017.  The date format requirement will be relaxed when I have time.  The first row should be Eclipse style names with the format Keyword:Name, as WBHP:PROD1.  Optionally, a DAYS column may be added.  DAYS is the number of days since the start of the simulatiion, and the default is NA.  The case name assigned to this data will be the basename of the input file, without the extent.
}
