% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plotting.R
\name{ploteach}
\alias{ploteach}
\title{Create individual plot for each name and keyword in each case}
\usage{
ploteach(long, casenames = NULL, wgnames = NULL, keywords = NULL,
  ncolumns = 3)
}
\arguments{
\item{long}{A long format summary dataframe}

\item{casenames}{A list with casenames to be plotted}

\item{wgnames}{A list with well/group/filed names to be plotted}

\item{keywords}{A list with parameters to be plotted, e.g. "WOPR"}

\item{ncolumns}{How many columns of plots to display}
}
\description{
This function accepts a long dataframe as input, and creates plots using ggplot
}
