% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preparing.R
\name{readDeck}
\alias{readDeck}
\title{Read in in data deck}
\usage{
readDeck(deckname = NULL, kw = "PORO")
}
\arguments{
\item{deckname}{The full path to, and name of, a simulation deck.  There is no default, and the function will faile without an input deck.}

\item{kw}{is the name of the property that one wishes to retrieve.}
}
\value{
Returns a data frame with the desired data.
}
\description{
This function reads selected data from a data deck into R
}
\details{
The current version of this function is very simple, and will read only a very limited selection of grid data.
}
