% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_hmsetup.R
\name{readtemplate}
\alias{readtemplate}
\title{Read a deck template, and create a template object.}
\usage{
readtemplate(template = NULL, basedir = ".")
}
\arguments{
\item{template}{The basename of an Eclipse style template deck.}

\item{basedir}{The path to the base directory of a simulation project.  The default is the current directory.}
}
\value{
Returns a template object, and writes out a template object file.
}
\description{
This function reads the variables defined in a template deck, and sets up an object with assumed distribution types and parameters.  It is the constructor for the "hmvars" class.
}
\details{
I'm thinking about it.
}
