% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_utils.R
\name{rss2d}
\alias{rss2d}
\title{A substitute for the DiceDesign::rss2d to fix a bug}
\usage{
rss2d(design, lower, upper, gof.test.type = "greenwood",
  gof.test.stat = NULL, transform = NULL, n.angle = 360, graphics = 1,
  trace = TRUE, lines.lwd = 1, lines.lty = "dotted", ...)
}
