% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_running.R
\name{runflow}
\alias{runflow}
\title{Run an Open Porous Media reservoir simulation}
\usage{
runflow(deckname, outdir = ".", sim_exec = "flow", restart = FALSE,
  sim_version = "stable", queue = "background")
}
\arguments{
\item{deckname}{The name and full path to an Eclipse style simulation input deck}

\item{outdir}{The base directory where the ouput should be written, assumed to be the current directory.}

\item{sim_exec}{one of ("flow", "flow_mpi", "flow_polymer", "flow_sequential", "flow_solvent")}

\item{restart}{Is this a restart run? (assumes FALSE)}

\item{sim_version}{one of c("stable", "latest")}

\item{queue}{one of c("background", "queue_name", "AWS").  Currently only "background" is functional.}
}
\value{
The function uses system2 to execute the simulation run.  Standard error and standard out are redirected to a file.  A non-zero return indicates an error. The function's purpose is the side effect of submitting a job asynchronously.
}
\description{
This function deals with the details of running a reservoir simulation model.  One may specify the model deck, the output location, the executable type and version, and the computing resources used to run the job.
}
\details{
A subdirectory of the the outdir with the base name of the input deck will be created, keeping each simulation run in its own directory.
Restarts probably need to be thought out a little better.
The stable executable is assumed to be in /usr/bin and the latest version in /usr/bin/local.  This is consistent with "stable" being the version installed from binaries and "latest" being compiled locally.
Only the "background" queue currently works.  Submission to a PBS queue is a short term goal and submission to an AWS batch queue is a longer term goal.
It might be desirable to look at the deckname file to determine which executable is appropriate (i.e. which keywords are in the deck).
As OPM simulators are not currently supported on Windows, this package is linux specific.
}
\references{
http://opm-project.org/
}
\author{
George Williams
}
