% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_runhm.R
\name{SelectModels}
\alias{SelectModels}
\title{SelectModels:  Help filter the member error data frame to select data for the desired kriged models}
\usage{
SelectModels(member_error = NULL, basedir = "tmp", wgnames = NULL,
  keywords = NULL, errortypes = NULL)
}
\arguments{
\item{member_error}{blah}

\item{basedir}{blah}

\item{wgnames}{blah}

\item{keywords}{blah}

\item{errortypes}{blah}
}
\value{
blah
}
\description{
blah
}
\details{
blah
}
